Following Best Practices
Looking at section 5.3 of baeldung's clean code they show an example that shows a well-formed source file as an example. This is also explained further in rhamedy's summary of best practices where it is shown with in line code comments the order of what should be expected in a code source file for readability. For these reasons I have restructured the file to have the private variables declared in order of visibility, followed by the base constructor and overloaded in sequential order, and at the bottom with the methods, getters, and setters needed for the project.
Sources:
https://rhamedy.medium.com/a-short-summary-of-java-coding-best-practices-31283d0167d3
https://www.baeldung.com/java-clean-code