Removal of Method
This method removed seemed to have an unused end goal. This was removed on the basis of unfocused code as described in the Characteristics of Clean Code described in baeldung. The result returns a null value and if it was used later in development it could serve potential unhandled nulls which would be harder to trace back to this unlogged null. With the code being unused and not serving current function it would be my recommendation to remove the method.
Source:
https://www.baeldung.com/java-clean-code